Overworld void generation, configuration, and more!
Sorry for the lack of updates, I was planning on releasing some fixes
and a few major updates all in this version.

Without further ado, here are the changes:
- Added mod configuration
- Added the option to generate a void overworld, disabled by default
(player will load in at height 77 with a single dirt block below them by
default, configuration coming soon)
- Biome generation has been reverted to vanilla generation, meaning that
mobs will be able to spawn in the overworld and nether, but can be
changed back to the void biome setting in the mod's configuration
- Cleaned up classes and packaging

I love hearing the feedback everyone! Thank you so much for all of the
downloads and support. Please keep the bug reports and suggestions
coming!