Void biomes with starting blocks progress towards 0.3
Huge thanks to Ferdz for this one, he helped me out with a huge issue I
was dealing with that was pushing back the development of the mod.

Here are the changes:
- Void biome overworld now has a starting block
- Restructured the mod and removed a lot of unnecessary code and
rearranged methods

This will open up the doors to a 0.3 release and means that a 1.0
release is coming very soon!