0.3
The time has finally come for a 0.3 release. Sorry everyone for the late
release. School and the holiday season has been crazy and I was faced
with a lot of challenges in pushing this update.

Huge thanks to /u/Ferdzz for this release. I couldn't have done it
without his help. He's helped me out a ton of my brain farts and has
helped me learn a lot about modding in Minecraft.

Here are the changes:
- Void biome overworld now has a starting block
- Restructured the mod and removed a lot of unnecessary code and
rearranged methods
- New modular style for mod classes that will allow for much easier
transition to 1.0 and beyond

Here is an overview of nVoid's features as of 0.3:
- Configurable spawn height
- Option to set the nether or the overworld to void chunk generation
- Option to set biome of the nether or the overworld to void