1.0
First release of nVoid with a nearly finished End. The end is able to
spawn as a void, the default obsidian platform. There is a toggle for
the end dragon.

At this point, there are a a bit too many features to list in the
changelog, so I invite everyone to check out the config file, mess
around with some settings and check out nVoid!

As always, please give me feedback on the mod and suggestions, and I'll
work towards squashing bugs and adding new features you all may desire!

Thanks and happy holidays!